requirements.txt
pymongo
openai
langchain>=0.1.13
langchain-community>=0.0.21
gradio
textblob
python-dotenv
tqdm